+++ CHANGES.md
@@ -1,3 +1,8 @@
+# Version 2.1.6 (to be released)
+
+* Fixed possible memory leak by removing decoration from layered pane when decorated component no longer in the component hierarchy tree
+* Fixed bug where IconComponentDecoration tooltip was constantly remaining on screen
+
 # Version 2.1.5
 
 * Added initial support for JavaFX 2